feat(ui): add home navigation on homeBtn click
